refactor: move injector hooks to 'inject'
